+++ packages/angular/src/app/modules/builder/utils/constants.ts
@@ -1,0 +1,3 @@
+// TODO remove hardcoded version, maybe a release tag?
+export const ANGULAR_LATEST_VERSION = '1.3.49';
+export const SCRIPT_ID = 'builder-wc-script';