+++ packages/sdks/src/blocks/symbol/symbol.lite.tsx
@@ -67,6 +67,7 @@ export default function Symbol(props: SymbolProps) {
       getContent({
         model: symbolToUse.model,
         apiKey: builderContext.apiKey,
+        apiVersion: builderContext.apiVersion,
         query: {
           id: symbolToUse.entry,
         },
@@ -87,6 +88,7 @@ export default function Symbol(props: SymbolProps) {
     >
       {JSON.stringify(state.contentToUse)}
       <RenderContent
+        apiVersion={builderContext.apiVersion}
         apiKey={builderContext.apiKey!}
         context={builderContext.context}
         customComponents={Object.values(builderContext.registeredComponents)}