+++ packages/sdks/src/blocks/text/text.lite.tsx
@@ -1,3 +1,9 @@
 export default function Text(props: { text: string }) {
-  return <span class="builder-text" innerHTML={props.text} />;
+  return (
+    <span
+      class="builder-text"
+      innerHTML={props.text}
+      style={{ outline: 'none' }}
+    />
+  );
 }