+++ packages/sdks/src/functions/get-content/generate-content-url.test.ts
@@ -1,4 +1,5 @@
 import { generateContentUrl } from './generate-content-url';
+import type { GetContentOptions } from './types';
 
 const testKey = 'YJIGb4i01jvw0SRdL5Bt';
 const testModel = 'page';
@@ -23,6 +24,7 @@ describe('Generate Content URL', () => {
     });
     expect(output).toMatchSnapshot();
   });
+
   test('Handles overrides correctly', () => {
     const output = generateContentUrl({
       apiKey: testKey,
@@ -31,5 +33,53 @@ describe('Generate Content URL', () => {
       options,
     });
     expect(output).toMatchSnapshot();
+  });
+
+  test('generate content url with apiVersion as v2', () => {
+    const output = generateContentUrl({
+      apiKey: testKey,
+      model: testModel,
+      query: { id: testId },
+      options,
+      apiVersion: 'v2',
+    });
+    expect(output).toMatchSnapshot();
+  });
+
+  test('generate content url with apiVersion as v3', () => {
+    const output = generateContentUrl({
+      apiKey: testKey,
+      model: testModel,
+      query: { id: testId },
+      options,
+      apiVersion: 'v3',
+    });
+    expect(output).toMatchSnapshot();
+  });
+
+  test('throw error when trying to generate content url with apiVersion as v1', () => {
+    expect(() => {
+      generateContentUrl({
+        apiKey: testKey,
+        model: testModel,
+        query: { id: testId },
+        options,
+        apiVersion: 'v1' as GetContentOptions['apiVersion'],
+      });
+    }).toThrow(`Invalid apiVersion: expected 'v2' or 'v3', received 'v1'`);
+  });
+
+  test('throw error when trying to generate content url with an invalid apiVersion value', () => {
+    expect(() => {
+      generateContentUrl({
+        apiKey: testKey,
+        model: testModel,
+        query: { id: testId },
+        options,
+        apiVersion: 'INVALID_API_VERSION' as GetContentOptions['apiVersion'],
+      });
+    }).toThrow(
+      `Invalid apiVersion: expected 'v2' or 'v3', received 'INVALID_API_VERSION'`
+    );
   });
 });